requirements.txt
torch>=2.2.0
torchvision>=0.17.0
triton>=3.0.0
transformers>=4.40.0
accelerate>=0.28.0
diffusers>=0.25.0
safetensors>=0.4.0
tokenizers>=0.15.0
opencv-python>=4.7.0
pillow>=10.0.0
timm>=0.9.0
einops>=0.7.0
scikit-image
numpy>=1.24
pandas>=2.0
scipy
tqdm
omegaconf
pyyaml
requests
sentencepiece